Sync temporary changes. Do not forget to ammend.
