+++ .vscode/settings.json
@@ -1,5 +1,5 @@
 {
-    "python.formatting.provider": "black",
+    "python.formatting.provider": "none",
     "python.testing.unittestEnabled": false,
     "python.testing.unittestArgs": [
         "-v",
@@ -25,5 +25,10 @@
         "tests",
         "-s"
     ],
-    "python.envFile": "${workspaceFolder}/.env"
+    "python.envFile": "${workspaceFolder}/.env",
+    "python.analysis.typeCheckingMode": "off",
+    "[python]": {
+        "editor.defaultFormatter": "ms-python.black-formatter"
+    },
+    "python.analysis.autoImportCompletions": false
 }
