+++ seemps/cross/__init__.py
@@ -1,2 +1,2 @@
-from .cross import Cross
+from .cross import Cross, run_cross
 from .mesh import *
